+++ example/slow.php
@@ -2,4 +2,4 @@
 header("Cache-Control: s-maxage=5");
 header("Date: " . date("D, j M Y G:i:s T"));
 sleep(1);
-echo "Hello, Woooooooooooorld<br>\n";
+echo "Hello, World<br>\n";